requirements.txt
fastapi==0.95
orjson==3.8.8
uvicorn==0.21.1
python-dotenv==1.0.0
SQLAlchemy==2.0.7
alembic==1.10.2
asyncpg==0.27.0
pydantic==1.10.6
python-multipart==0.0.6
pytest==7.2.2
pytest_asyncio==0.21.0
httpx==0.23.3
starlette==0.26.1

